Move Consumption structures to parse/consume.rs
