+++ src/parse/rule/collect/mod.rs
@@ -25,9 +25,9 @@
 
 mod prelude {
     pub use super::try_collect;
-    pub use crate::parse::consume::consume;
+    pub use crate::parse::consume::{consume, Consumption, GenericConsumption};
     pub use crate::parse::error::{ParseError, ParseErrorKind};
-    pub use crate::parse::rule::{Consumption, GenericConsumption, Rule};
+    pub use crate::parse::rule::Rule;
     pub use crate::parse::token::{ExtractedToken, Token};
     pub use crate::text::FullText;
 }
