+++ src/parse/rule/impls/mod.rs
@@ -19,10 +19,10 @@
  */
 
 mod prelude {
-    pub use crate::parse::consume::consume;
+    pub use crate::parse::consume::{consume, Consumption, GenericConsumption};
     pub use crate::parse::error::{ParseError, ParseErrorKind, ParseException};
     pub use crate::parse::rule::collect::*;
-    pub use crate::parse::rule::{Consumption, GenericConsumption, Rule, TryConsumeFn};
+    pub use crate::parse::rule::{Rule, TryConsumeFn};
     pub use crate::parse::token::{ExtractedToken, Token};
     pub use crate::text::FullText;
     pub use crate::tree::{Container, ContainerType, Element};
